// src/shared/types/serverTypes.ts
export type Category = {
  id: string;
  name: string;
  photo?: string;
  createdAt: Date;
  updatedAt: Date;
  commandId: string;
};
// ...
export type Pagination = {
  pageSize?: number;
  pageNumber?: number;
};
// ...
export type MutateCategoryBody = Omit<Category, 'id' | 'createdAt' | 'updatedAt' | 'commandId'>;